Extract theme region into sections/theme.js
P-08 of conception/projects/2026-04-23-condash-frontend-extraction. First of six planned region extractions in this PR.

Theme is the simplest of the P-08 candidates: 40 LOC, one exported side-effect (the initial applyTheme call, now wrapped in initThemeSideEffects), and two cross-region references (_cmViews + _currentCmTheme) that only fire inside applyTheme's function body — safe under the same circular-import rules P-07 established.

Bundle builds clean with make frontend.
